(refactor) Consolidate object storage test helpers into `common` module and use RustFS as default container
