+++ crates/icegate-maintain/tests/iceberg_source_it.rs
@@ -19,6 +19,7 @@
 
 #![allow(clippy::expect_used, clippy::unwrap_used, clippy::too_many_lines)]
 
+mod common;
 use std::cmp::Ordering;
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -27,8 +28,8 @@ use arrow::array::{Array, FixedSizeBinaryArray, MapArray, StringArray, StructArr
 use arrow::buffer::OffsetBuffer;
 use arrow::datatypes::{DataType, Schema as ArrowSchema};
 use arrow::record_batch::RecordBatch;
+use common::{build_s3_catalog, setup_object_store};
 use futures::TryStreamExt;
-use iceberg::io::FileIOBuilder;
 use iceberg::spec::{DataFile, DataFileFormat};
 use iceberg::table::Table;
 use iceberg::transaction::{ApplyTransactionAction, Transaction};
@@ -39,75 +40,22 @@ use iceberg::writer::file_writer::rolling_writer::RollingFileWriterBuilder;
 use iceberg::writer::partitioning::PartitioningWriter;
 use iceberg::writer::partitioning::fanout_writer::FanoutWriter;
 use iceberg::{Catalog, NamespaceIdent, TableCreation, TableIdent};
-use icegate_catalog_s3::{CatalogCodecKind, S3Catalog, S3CatalogConfig};
-use icegate_common::catalog::IoHandle;
+use icegate_catalog_s3::S3Catalog;
 use icegate_common::manifest_scan::{DataFileStats, list_data_files_with_stats};
 use icegate_common::merge::sort_key::{SortColumnCache, SortColumnsDescriptor};
 use icegate_common::merge::{MergeInput, MergePosition, RowGroupsMerger, SortedBatchMergerConfig};
 use icegate_common::schema::{logs_partition_spec, logs_schema, logs_sort_order};
-use icegate_common::testing::{S3TestContainer, create_s3_bucket};
 use icegate_maintain::compact::iceberg_source::IcebergMergeSource;
 use parquet::file::properties::WriterProperties;
 use tokio_util::sync::CancellationToken;
 use uuid::Uuid;
 
-const BUCKET_NAME: &str = "warehouse";
 const NAMESPACE: &str = "icegate";
 const TABLE: &str = "logs";
 const TENANT: &str = "tenant-a";
 /// 2026-06-11T00:00:00Z in microseconds. Both seeded files share this day so
 /// they land in the same `(tenant_id, day)` partition.
 const DAY_MICROS: i64 = 1_749_600_000_000_000;
-
-/// Connection parameters for a running object store.
-#[derive(Clone)]
-struct StorageConn {
-    endpoint: String,
-    access_key: String,
-    secret_key: String,
-}
-
-/// Stand up object storage and capture its connection parameters.
-async fn setup_object_store() -> (S3TestContainer, StorageConn) {
-    let store = S3TestContainer::start().await.expect("start object storage");
-    create_s3_bucket(store.endpoint(), BUCKET_NAME).await.expect("create bucket");
-    let conn = StorageConn {
-        endpoint: store.endpoint().to_string(),
-        access_key: store.username().to_string(),
-        secret_key: store.password().to_string(),
-    };
-    (store, conn)
-}
-
-/// Build a concrete [`S3Catalog`] against the object store.
-async fn build_s3_catalog(conn: &StorageConn) -> S3Catalog {
-    let io = IoHandle::noop();
-    let mut props: HashMap<String, String> = HashMap::new();
-    props.insert("warehouse".to_string(), format!("s3://{BUCKET_NAME}"));
-    props.insert("s3.endpoint".to_string(), conn.endpoint.clone());
-    props.insert("s3.path-style-access".to_string(), "true".to_string());
-    props.insert("s3.access-key-id".to_string(), conn.access_key.clone());
-    props.insert("s3.secret-access-key".to_string(), conn.secret_key.clone());
-    props.insert("s3.region".to_string(), "us-east-1".to_string());
-    let file_io = FileIOBuilder::new(io.storage_factory()).with_props(props).build();
-
-    S3Catalog::new(
-        S3CatalogConfig {
-            bucket: BUCKET_NAME.to_string(),
-            region: "us-east-1".to_string(),
-            endpoint: Some(conn.endpoint.clone()),
-            access_key_id: Some(conn.access_key.clone()),
-            secret_access_key: Some(conn.secret_key.clone()),
-            warehouse: BUCKET_NAME.to_string(),
-            codec: CatalogCodecKind::Json,
-            ..S3CatalogConfig::default()
-        },
-        file_io,
-        tokio_util::sync::CancellationToken::new(),
-    )
-    .await
-    .expect("build S3 catalog")
-}
 
 /// Create the namespace and `logs` table, returning the table identifier.
 async fn create_logs_table(catalog: &S3Catalog) -> TableIdent {
